replace sqlite/postgres with H2
