+++ src/main/java/org/peergos/EmbeddedIpfs.java
@@ -8,7 +8,10 @@ import io.libp2p.core.multistream.*;
 import io.libp2p.protocol.*;
 import org.peergos.blockstore.*;
 import org.peergos.blockstore.metadatadb.BlockMetadataStore;
+import org.peergos.blockstore.metadatadb.JdbcBlockMetadataStore;
 import org.peergos.blockstore.metadatadb.MetadataCachingStorage;
+import org.peergos.blockstore.metadatadb.sql.SqliteCommands;
+import org.peergos.blockstore.metadatadb.sql.UncloseableConnection;
 import org.peergos.blockstore.s3.S3Blockstore;
 import org.peergos.config.*;
 import org.peergos.net.ConnectionException;
@@ -21,6 +24,9 @@ import org.peergos.protocol.http.*;
 import org.peergos.util.Logging;
 
 import java.nio.file.*;
+import java.sql.Connection;
+import java.sql.DriverManager;
+import java.sql.SQLException;
 import java.util.*;
 import java.util.concurrent.*;
 import java.util.logging.*;
@@ -115,6 +121,20 @@ public class EmbeddedIpfs {
         return node != null ? node.stop() : CompletableFuture.completedFuture(null);
     }
 
+    public static BlockMetadataStore buildBlockMetadata(Args a) {
+        try {
+            //see https://www.h2database.com/html/features.html#compatibility for the extra params to support
+            // compatibility mode. This is required for 'ON CONFLICT DO NOTHING aka INSERT OR IGNORE INTO'
+            Path metadataPath = a.fromIPFSDir("nabu-block-metadata-sql-file", "nabu-blockmetadata.sql");
+            java.sql.Connection h2Instance = DriverManager.getConnection("jdbc:h2:" +
+                    metadataPath.toString() + ";MODE=PostgreSQL;DATABASE_TO_LOWER=TRUE;DEFAULT_NULL_ORDERING=HIGH");
+            Connection instance = new UncloseableConnection(h2Instance);
+            instance.setAutoCommit(true);
+            return new JdbcBlockMetadataStore(() -> instance, new SqliteCommands());
+        } catch (SQLException e) {
+            throw new RuntimeException(e);
+        }
+    }
     public static Blockstore buildBlockStore(Config config, Path ipfsPath, BlockMetadataStore meta) {
         Blockstore blocks = null;
         if (config.datastore.blockMount.prefix.equals("flatfs.datastore")) {
